requirements.txt
pytest==7.1.2
selenium==4.1.0
requests==2.27.1
pytest-xdist==2.5.0
allure-pytest==2.9.45
